+++ src/ui/activities/setup_activity/mod.rs
@@ -54,6 +54,7 @@ enum UserInterfaceInputField {
     DefaultProtocol,
     TextEditor,
     ShowHiddenFiles,
+    CheckForUpdates,
     GroupDirs,
     FileFmt,
 }