+++ CHANGELOG.md
@@ -17,9 +17,12 @@
 
 Released on 28/02/2021
 
-- Format key attributes:
+- **Format key attributes**:
   - Added `EXTRA` and `LENGTH` parameters to format keys.
   - Now keys are provided with this syntax `{KEY_NAME[:LEN[:EXTRA]}`
+- **Check for updates**:
+  - TermSCP will now check for updates on startup and will show in the main page if there is a new version available
+  - This feature may be disabled from setup (Check for updates => No)
 - Enhancements:
   - Default choice for deleting file set to "NO" (way too easy to delete files by mistake)
   - Added CLI options to set starting workind directory on both local and remote hosts